+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 - **`@to_timestamp`**: Child decorator to convert a `date`, `time`, or `datetime` object to a unix timestamp.
 - **`@add_prefix`**: Child decorator to add a prefix to a string.
 - **`@add_suffix`**: Child decorator to add a suffix to a string.
+- **`@remove_prefix`**: Child decorator to remove a prefix from a string.
+- **`@remove_suffix`**: Child decorator to remove a suffix from a string.
 - **`@to_date`**: Child decorator to convert a string from various formats to a `date` object.
 - **`@to_time`**: Child decorator to convert a string from various formats to a `time` object.
 - **`@to_datetime`**: Child decorator to convert a string from various formats to a `datetime` object.